pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "focusforge"
version = "0.1.0"
description = "Agentic CLI productivity tool using Python, LangChain, GPT, and Spotify integration."
authors = [{ name = "Douglas Danso"}]
dependencies = [
    "openai",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "python-dotenv",
    "spotipy",
    "rich",
    "click",
    "pydantic",
    "tqdm"
]

[tool.hatch.envs.default]
dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["focusforge"]